% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{assign_csvs_with_prefix}
\alias{assign_csvs_with_prefix}
\title{Util function for loading a directory of csvs
and assigning the resulting data frames with
prefixed names.}
\usage{
assign_csvs_with_prefix(directory, prefix)
}
\description{
Util function for loading a directory of csvs
and assigning the resulting data frames with
prefixed names.
}
