% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_final_score}
\alias{filter_final_score}
\title{Select only the rows containing the team's highest score.}
\usage{
filter_final_score(frame)
}
\description{
Select only the rows containing the team's highest score.
}

