% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totems-data.R
\docType{data}
\name{Inventories}
\alias{Inventories}
\title{Inventories uncovered by teams in the Totems experiment.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 928 rows and 7 columns.}
\usage{
data("Inventories")
}
\description{
\describe{
  \item{PlayerID}{unique identifier for players}
  \item{NumInnovations}{number of new items discovered by the team}
  \item{Inventory}{a string representing the current items in the inventory}
}
}
\keyword{datasets}
