% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totems-data.R
\docType{data}
\name{PlayerInfo}
\alias{PlayerInfo}
\title{Information about players}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 146 rows and 4 columns.}
\usage{
data("PlayerInfo")
}
\description{
Information about players
}
\keyword{datasets}
